requirements.txt
pymongo==4.8.0
neo4j==5.23.0
mysql-connector-python==9.0.0
dash==2.17.1
dash-bootstrap-components==1.6.0
pandas==2.2.2